pyproject.toml
[project]
name = "pokemon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.122.0",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
]
